Add OCI IMDS v2 support
Instance Metadata (IMDS) is a service provided to all overlay instances* in OCI
and is utilized by many OCI services and external customers. It responds to
HTTP requests from services running on the instance via the 169.254.169.254
interface and serves up metadata about the instance such as region info,
networking details, customer-provided metadata, Identity certificates, and many
others.  In an effort to increase the security of IMDS we will be providing a
new /opc/v2 endpoint for OCI services and customers to use which requires a
security header to be added to all requests (Authorization: Bearer Oracle) and
will additionally reject forwarded requests (requests which include HTTP header
Forwarded and/or X-Forwarded-For).  Any requests to /opc/v2 which don't provide
this header will be rejected (403).  The /opc/v2 endpoint will be enabled by
default for all instances.

https://confluence.oci.oraclecorp.com/display/C3E/Instance+Metadata%3A+Updating+Your+Code+to+Support+IMDS+v2

The patch adds the support to visit the /opc/v2 endpoint with proper request
header - "Authorization: Bearer Oracle", and if v2 is not available it will
fall back to v1.

Signed-off-by: Frank Deng <frank.deng@oracle.com>
Reviewed-by: Akira Ijuin <akira.ijuin@oracle.com>